Commit for Selects
Modified best_in_place.js for select boxes.  Basically, it performs a
RegEx check on the displayed text prior to creating the select field.

The case is as follows: If the value in the collection is something
like [[1, "Private"],[2, "Public"]], but :display_as renders the field
as "Private - Some extra data," it performs an additional RegEx check
for value[1] (in this case, "Private"), plus " - ".  Allows you to put
descriptors, and have a smaller select field.