Added tests, fixed the js
Added tests to make sure it works and also that :type => :date still
works as well.

I accidently changed the "date" in the best_in_pace.js to use datetime
so I changed that back.

jQuery-ui now comes from a Gem - not using the CSS

The slider doesn't work in this app - not sure why, could be the
ancient CSS.